requirements.txt
pandas
supabase
streamlit
anthropic
google-generativeai
requests
